refactoring routes contacts
