+++ routes/contacts/contacts.js
@@ -2,29 +2,32 @@ const express = require("express");
 const router = express.Router();
 
 const {
+  validateContactId,
   validateContact,
   validateContactPatch,
-  validateContactId,
 } = require("./validation");
 
-router.get("/", async (req, res, next) => {
-  res.json({ message: "template message" });
-});
+const {
+  getContacts,
+  getContact,
+  addContact,
+  updateContact,
+  deleteContact,
+} = require("../../controllers/contacts");
 
-router.get("/:contactId", async (req, res, next) => {
-  res.json({ message: "template message" });
-});
+router.get("/", getContacts);
 
-router.post("/", async (req, res, next) => {
-  res.json({ message: "template message" });
-});
+router.get("/:contactId", validateContactId, getContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-  res.json({ message: "template message" });
-});
+router.post("/", validateContact, addContact);
 
-router.patch("/:contactId", async (req, res, next) => {
-  res.json({ message: "template message" });
-});
+router.delete("/:contactId", validateContactId, deleteContact);
+router.put("/:contactId", validateContactId, validateContact, updateContact);
+router.patch(
+  "/:contactId",
+  validateContactId,
+  validateContactPatch,
+  updateContact
+);
 
 module.exports = router;
